+++ Sources/TBRESTClientLib/TbApiEndpoints.swift
@@ -12,6 +12,7 @@ protocol SupportedTbAPIEndpoints {
     var getCustomerById: String { get }
     var getCustomerDevices: String { get }
     var getCustomerDeviceInfos: String { get }
+    var getTenantDevices: String { get }
     var getDeviceById: String { get }
     var getDeviceInfoById: String { get }
     var saveDevice: String { get }
@@ -43,6 +44,7 @@ struct TbAPIEndpointsV1: SupportedTbAPIEndpoints {
     public let getCustomerById = "/api/customer/{?customerId?}"
     public let getCustomerDevices = "/api/customer/{?customerId?}/devices"
     public let getCustomerDeviceInfos = "/api/customer/{?customerId?}/deviceInfos"
+    public let getTenantDevices = "/api/tenant/devices"
     public let getDeviceById = "/api/device/{?deviceId?}"
     public let getDeviceInfoById = "/api/device/info/{?deviceId?}"
     public let saveDevice = "/api/device{?accessToken?}"
@@ -72,6 +74,7 @@ struct TbAPIEndpointsV2: SupportedTbAPIEndpoints {
     public let getCustomerById = ""
     public let getCustomerDevices = ""
     public let getCustomerDeviceInfos = ""
+    public let getTenantDevices = ""
     public let getDeviceById = ""
     public let getDeviceInfoById = ""
     public let saveDevice = ""